+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.0.2] - 05-12-2019
+
 ### Added
 
 - `Heading` component
@@ -38,5 +40,7 @@ and this project adheres to
 - `PageAnnouncement` component
 
 [unreleased]:
-  https://github.com/LBHackney-IT/lbh-frontend-react/compare/v0.0.1...HEAD
+  https://github.com/LBHackney-IT/lbh-frontend-react/compare/v0.0.2...HEAD
+[0.0.2]:
+  https://github.com/LBHackney-IT/lbh-frontend-react/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/LBHackney-IT/lbh-frontend-react/releases/tag/v0.0.1